rf: renames, change components compositon
